refactor: update artist selection components to use search params and improve query handling
